// packages/lightspeed/src/interfaces.ts
export interface CompletionResponseParams {
    predictions: string[];
    model?: string;
    suggestionId: string;
}

export interface MetadataParams {
    documentUri: string;
    ansibleFileType: IAnsibleFileType;
    activityId: string;
    additionalContext?: IAdditionalContext;
}

export interface CompletionRequestParams {
    prompt: string;
    suggestionId?: string;
    metadata?: MetadataParams;
    model?: string;
}

export interface FeedbackResponseParams {
    message?: string;
}

export interface InlineSuggestionEvent {
    latency?: number;
    userActionTime?: number;
    documentUri?: string;
    action?: UserAction;
    error?: string;
    suggestionId?: string;
    activityId?: string;
}

export interface PlaybookExplanationEvent {
    explanationId?: string;
}

export interface PlaybookFeedbackEvent {
    action: ThumbsUpDownAction;
    explanationId?: string;
    generationId?: string;
}

export interface PlaybookGenerationActionEvent {
    wizardId: string;
    action: WizardGenerationActionType;
    fromPage?: number;
    toPage?: number;
}

export interface RoleGenerationActionEvent {
    wizardId: string;
    action: WizardGenerationActionType;
    fromPage?: number;
    toPage?: number;
}

export interface RoleExplanationEvent {
    explanationId?: string;
}

export interface RoleFeedbackEvent {
    action: ThumbsUpDownAction;
    explanationId?: string;
    generationId?: string;
}

export interface FeedbackRequestParams {
    inlineSuggestion?: InlineSuggestionEvent;
    playbookExplanation?: PlaybookExplanationEvent;
    playbookExplanationFeedback?: PlaybookFeedbackEvent;
    playbookGenerationAction?: PlaybookGenerationActionEvent;
    roleGenerationAction?: RoleGenerationActionEvent;
    roleExplanation?: RoleExplanationEvent;
    roleExplanationFeedback?: RoleFeedbackEvent;
    playbookOutlineFeedback?: PlaybookFeedbackEvent;
    model?: string;
}

export interface IDocumentTrackerFields {
    activityId: string;
    content: string;
}

export type IDocumentTracker = Record<string, IDocumentTrackerFields>;

export interface ContentMatchesRequestParams {
    suggestions: string[];
    suggestionId: string;
    model?: string;
}

export interface IContentMatchParams {
    repo_name: string;
    repo_url: string;
    path: string;
    license: string;
    data_source_description: string;
    score: number;
}

export interface IContentMatch {
    contentmatch: IContentMatchParams[];
}

export interface ContentMatchesResponseParams {
    contentmatches: IContentMatch[];
}

export interface ISuggestionDetails {
    suggestion: string;
    suggestionId: string;
    isPlaybook: boolean;
}

export interface PlaybookGenerationRequestParams {
    text: string;
    outline?: string;
    generationId: string;
    createOutline: boolean;
    wizardId?: string;
}

export interface RoleGenerationRequestParams {
    name?: string;
    text: string;
    outline?: string;
    generationId: string;
    createOutline: boolean;
    wizardId?: string;
}

export interface PlaybookGenerationResponseParams {
    playbook: string;
    outline?: string;
    generationId: string;
}

export enum RoleFileType {
    Default = 'default',
    Task = 'task',
    Playbook = 'playbook',
    Handler = 'handler',
}

export interface GenerationListEntry {
    path: string;
    file_type: RoleFileType;
    content: string;
}

export interface RoleGenerationResponseParams {
    files: GenerationListEntry[];
    outline?: string;
    generationId: string;
    name: string;
    role?: string;
}

export interface RoleExplanationRequestParams {
    files: GenerationListEntry[];
    explanationId: string;
    roleName: string;
}

export interface ExplanationRequestParams {
    content: string;
    explanationId: string;
}

export interface ExplanationResponseParams {
    content: string;
    format: string;
    explanationId: string;
}

export type IAnsibleFileType = 'playbook' | 'tasks_in_role' | 'tasks' | 'other';
// ...
interface IAdditionalContext {
    playbookContext?: IPlaybookContext;
    roleContext?: IRoleContext;
    standaloneTaskContext?: IStandaloneTaskContext;
}